files separated for trip itinerary
