+++ frontend/app/(app)/trips/[id]/components/itinerary-empty-state.tsx
@@ -1,0 +1,30 @@
+import { Box, Text } from "@/design-system";
+import { StyleSheet } from "react-native";
+
+// ─── Component ────────────────────────────────────────────────────────────────
+
+export function ItineraryEmptyState() {
+  return (
+    <Box borderWidth={1} borderColor="gray200" borderRadius="xl" padding="sm">
+      <Box alignItems="center" paddingVertical="lg">
+        <Text
+          variant="bodySmDefault"
+          color="gray950"
+          style={styles.emptyStateText}
+        >
+          No activites planned. Get planning!
+        </Text>
+      </Box>
+    </Box>
+  );
+}
+
+// ─── Styles ───────────────────────────────────────────────────────────────────
+
+const styles = StyleSheet.create({
+  emptyStateText: {
+    fontStyle: "italic",
+  },
+});
+
+export default ItineraryEmptyState;
